+++ CHANGES.rst
@@ -16,6 +16,8 @@ The Products.MeetingCommunes version must be the same as the Products.PloneMeeti
   [gbastien]
 - Call PloneMeeting migrations to 4208 in MC migration to 4200.
   [gbastien]
+- Call PloneMeeting migrations to 4209 in MC migration to 4200.
+  [gbastien]
 
 4.2.3 (2023-07-07)
 ------------------