+++ CHANGES.rst
@@ -5,7 +5,7 @@ Products.MeetingCommunes Changelog
 The Products.MeetingCommunes version must be the same as the Products.PloneMeeting version
 
 
-4.2.4 (unreleased)
+4.2.4 (2023-09-12)
 ------------------
 
 - Updated `attendance-stats.ods`.